Switch to System.IO.Hashing.XxHash64 for file hash #76

+++ src/DbLive.MSSQL/SQL/03.unit-tests.sql
@@ -1,7 +1,7 @@
 ﻿
 create table dblive.unit_test_run (
 	relative_path nvarchar(4000) not null
-  , content_hash int not null
+  , content_hash bigint not null
   , run_utc datetime2(7) not null
   , execution_time_ms int not null
   , pass bit not null
